refactor(api): prevent type duplication user
